remove picture

+++ backend/app/controllers/api/user_spot_links_controller.rb
@@ -5,19 +5,26 @@ class Api::UserSpotLinksController < Api::BaseController
       @spot           = @mission.spots.find(params[:spot_id])
       @user_spot_link = @spot.user_spot_links.where(:user_id => @user.id).first_or_initialize
 
-      tmp_path = "#{Rails.root}/tmp/#{UUIDTools::UUID.random_create}.jpg"
-      data     = Base64.decode64(params[:user_spot_link][:picture])
+      if params[:user_spot_link][:picture]
+        tmp_path = "#{Rails.root}/tmp/#{UUIDTools::UUID.random_create}.jpg"
+        data     = Base64.decode64(params[:user_spot_link][:picture])
 
-      Rails.logger.info tmp_path
+        Rails.logger.info tmp_path
 
-      File.open(tmp_path, 'wb') do |file|
-        file.write(data)
+        File.open(tmp_path, 'wb') do |file|
+          file.write(data)
+        end
+
+        @user_spot_link.picture = File.new(tmp_path)
+      else
+        @user_spot_link.remove_picture!
       end
 
-      @user_spot_link.picture = File.new(tmp_path)
       @user_spot_link.save!
 
-      FileUtils.rm(tmp_path)
+      if tmp_path
+        FileUtils.rm(tmp_path)
+      end
     end
 
     render :nothing => true
